feat: 计时

+++ urban-management-pojo/src/main/java/com/unicom/urban/management/pojo/vo/EventVO.java
@@ -72,7 +72,7 @@ public class EventVO {
     /**
      * 警示灯 时长
      */
-    private Long timeLimitLong;
+    private int timeLimitLong;
 
     /**
      * 时限单位
